Moved hardware/stirring/__init__.py to hardware/stirring/core.py
